% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_diagonal.R
\name{is.diagonal}
\alias{is.diagonal}
\title{Test if matrix is diagonal}
\usage{
is.diagonal(x, na.rm = FALSE)
}
\arguments{
\item{x}{A matrix}

\item{na.rm}{TRUE/FALSE  TRUE means that NAs on the diagonal are ignored}
}
\description{
0s on the diagonal are fine.
}
\keyword{internal}
